cli: extract runtime/ subpackage (setup + report + run)
